+++ .gitignore
@@ -10,7 +10,7 @@
 /test/version_tmp/
 /tmp/
 /node_modules/
-/output/*
+
 
 # Used by dotenv library to load environment variables.
 # .env
